Talon SRX/FX motion magic v/a should be doubles

+++ core_robot/src/main/java/frc/robot/common/robotprovider/ITalonXBase.java
@@ -13,7 +13,7 @@ public interface ITalonXBase extends IMotor
     void configureAllowableClosedloopError(int slotId, int error);
     void setSelectedSlot(int slotId);
     void setPIDF(double p, double i, double d, double f, int slotId);
-    void setMotionMagicPIDF(double p, double i, double d, double f, int velocity, int acceleration, int slotId);
+    void setMotionMagicPIDF(double p, double i, double d, double f, double velocity, double acceleration, int slotId);
     void setPIDF(double p, double i, double d, double f, int izone, double closeLoopRampRate, int slotId);
     void setForwardLimitSwitch(boolean enabled, boolean normallyOpen);
     void setReverseLimitSwitch(boolean enabled, boolean normallyOpen);
